requirement.txt
streamlit
openai
pyPDF2
google-generativeai
